requirements.txt
numpy<1.22.3
opencv_python==4.5.5.64
Pillow==9.1.1
requests==2.27.1
streamlit==1.9.0
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
opencv-python-headless==4.5.5.64

